Add deliver_sm with optional parameters
The short_message is currently interpreted as a string directly from the bytes... I'm thinking of either letting it as an array of bytes so we can come and decode it later or leave it as string because strings in go are a sequence of bytes... The danger is that we do not fix the string after or that we should fix it before.

Ref.
https://blog.golang.org/strings
https://stackoverflow.com/questions/13510458/golang-convert-iso8859-1-to-utf8