Add networking of client and server (TCP)
The commit evolve the network test to handle errors and send a SMPP object into the write.  We handle the reading as well as the bytes (not a SMPP object).

This test will have to evolve to be more user-friendly and require to pursue the pdu_builder for submit_sm.