test 2 connections on the same smsc
This is far from the ideal solution as we're still in the same "thread'
of logical execution.  However, it added 2 new things for the tests;
- an exposed stucture for the connections
- a way to read from specific connections

This will probably evolve to have the SMSC have it's own co-routines
started to accept new incoming connections automatically.  It should
probably also get to a point where we should be able to get the errors
encountered in the SMSC co-routine... Problem for futur me/contributor.