Make Accepting connection implicit (not finished)
So we're moving toward a concurrent model.  Currently, the Accept()
is done explicitly, but we can start a goroutine to have it done in
background/"other thread".  When the "AcceptAllNewConnection" is
called, any new TCP syn sent will be added into our server connection
pool.  At this point, we would need to be able to engage with whatever
comes onto that queue and have the main execution thread be able to
interact with it as well.

I also had to start using pointers everywhere.  Other wise, I was
getting a copy of the "smsc" object for which the number of connections
wasn't adjusting to the real number behind.