Address concurrent read/write of connections on smsc
The smsc object has to be concurrent to support different
connections and handling of messages on those same
connections.  The first thing is that we need to let the smsc
execute its logic while also let the user interact with it.

This means we have multiple interaction points to plan;
- The main thread spawning and taking actions
- The smsc thread managing the accept/read calls
- The smsc thread toward the users with registered functions

The last one is concerning as we might not have all the controls
we want later to change it accordingly... Food for thoughts