Apply an alive channel for the state control loop
This enable the detection of the state control loop
aliveness which enable us to know if it's stopped or not.
If we do not do this, the concurrency on the sending and
closing of the loop keeps us in a deadlock when asking to
close the controlloop again (push (send) vs pull (receive)).
If the control loop as a way to emit (generate) information,
then we can pull that information.  In golang, the receiver channel
will not panic if the channel is closed, and we can know if the
channel is closed thanks to the second signature of channel.

Ref.
https://go101.org/article/channel.html